Handle case where there's no stored pad URL, consolidate config
